requirements.txt
fastapi
uvicorn
Pillow
numpy
python-multipart

opencv-python-headless
imagehash
slowapi
